backend/pyproject.toml
[tool.poetry]
name = "todo-backend-api"
version = "0.1.0"
description = "Backend API for Todo Full-Stack Web Application"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlmodel = "^0.0.16"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
uvicorn = "^0.24.0"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]